+++ RedPandaIDE/project.cpp
@@ -1597,6 +1597,7 @@ void Project::setName(const QString &newName)
     if (newName != mName) {
         mName = newName;
         mNode->text = newName;
+        setModified(true);
     }
 }
 
@@ -1883,7 +1884,7 @@ Qt::ItemFlags ProjectModel::flags(const QModelIndex &index) const
     if (!p)
         return Qt::NoItemFlags;
     if (p==mProject->node().get())
-        return Qt::ItemIsEnabled | Qt::ItemIsDropEnabled;
+        return Qt::ItemIsEnabled | Qt::ItemIsDropEnabled | Qt::ItemIsEditable;
     Qt::ItemFlags flags = Qt::ItemIsEnabled | Qt::ItemIsEditable | Qt::ItemIsSelectable | Qt::ItemIsDragEnabled;
     if (p->unitIndex<0) {
         flags.setFlag(Qt::ItemIsDropEnabled);
@@ -1901,8 +1902,13 @@ bool ProjectModel::setData(const QModelIndex &index, const QVariant &value, int 
     if (!node)
         return false;
     if (role == Qt::EditRole) {
-        if (node == mProject->node())
-            return false;
+        if (node == mProject->node()) {
+            QString newName = value.toString().trimmed();
+            if (newName.isEmpty())
+                return false;
+            mProject->setName(newName);
+            return true;
+        }
         int idx = node->unitIndex;
         if (idx >= 0) {
             //change unit name